+++ R/cpBody.R
@@ -59,6 +59,25 @@ cpBodySetPosition <- function(body, pos) {
 
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+#' Get the offset of the center of gravity in body local coordinates.
+#'
+#' Get the offset of the center of gravity in body local coordinates.
+#'
+#' C function prototype: \code{CP_EXPORT cpVect cpBodyGetCenterOfGravity(const cpBody *body);}
+#'
+#' @param body [\code{cpBody *}]
+#'
+#' @return [\code{cpVect *}]
+#'
+#' @export
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+cpBodyGetCenterOfGravity <- function(body) {
+  stopifnot(inherits(body, "cpBody"))
+  .Call("cpBodyGetCenterOfGravity_", body)
+}
+
+
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 #' Get the velocity of the body.
 #'
 #' Get the velocity of the body.
@@ -94,4 +113,118 @@ cpBodySetVelocity <- function(body, velocity) {
   stopifnot(inherits(velocity, "cpVect"))
   .Call("cpBodySetVelocity_", body, velocity)
   invisible(NULL)
+}
+
+
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+#' Get the force applied to the body for the next time step.
+#'
+#' Get the force applied to the body for the next time step.
+#'
+#' C function prototype: \code{CP_EXPORT cpVect cpBodyGetForce(const cpBody *body);}
+#'
+#' @param body [\code{cpBody *}]
+#'
+#' @return [\code{cpVect *}]
+#'
+#' @export
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+cpBodyGetForce <- function(body) {
+  stopifnot(inherits(body, "cpBody"))
+  .Call("cpBodyGetForce_", body)
+}
+
+
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+#' Get the angle of the body.
+#'
+#' Get the angle of the body.
+#'
+#' C function prototype: \code{CP_EXPORT cpFloat cpBodyGetAngle(const cpBody *body);}
+#'
+#' @param body [\code{cpBody *}]
+#'
+#' @return [\code{cpFloat}]
+#'
+#' @export
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+cpBodyGetAngle <- function(body) {
+  stopifnot(inherits(body, "cpBody"))
+  .Call("cpBodyGetAngle_", body)
+}
+
+
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+#' Set the angle of a body.
+#'
+#' Set the angle of a body.
+#'
+#' C function prototype: \code{CP_EXPORT void cpBodySetAngle(cpBody *body, cpFloat a);}
+#'
+#' @param body [\code{cpBody *}]
+#' @param a [\code{cpFloat}]
+#'
+#' @export
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+cpBodySetAngle <- function(body, a) {
+  stopifnot(inherits(body, "cpBody"))
+  .Call("cpBodySetAngle_", body, a)
+  invisible(NULL)
+}
+
+
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+#' Set the angular velocity of the body.
+#'
+#' Set the angular velocity of the body.
+#'
+#' C function prototype: \code{CP_EXPORT void cpBodySetAngularVelocity(cpBody *body, cpFloat angularVelocity);}
+#'
+#' @param body [\code{cpBody *}]
+#' @param angularVelocity [\code{cpFloat}]
+#'
+#' @export
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+cpBodySetAngularVelocity <- function(body, angularVelocity) {
+  stopifnot(inherits(body, "cpBody"))
+  .Call("cpBodySetAngularVelocity_", body, angularVelocity)
+  invisible(NULL)
+}
+
+
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+#' Set the torque applied to the body for the next time step.
+#'
+#' Set the torque applied to the body for the next time step.
+#'
+#' C function prototype: \code{CP_EXPORT void cpBodySetTorque(cpBody *body, cpFloat torque);}
+#'
+#' @param body [\code{cpBody *}]
+#' @param torque [\code{cpFloat}]
+#'
+#' @export
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+cpBodySetTorque <- function(body, torque) {
+  stopifnot(inherits(body, "cpBody"))
+  .Call("cpBodySetTorque_", body, torque)
+  invisible(NULL)
+}
+
+
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+#' Get the rotation vector of the body. (The x basis vector of it's transform.)
+#'
+#' Get the rotation vector of the body. (The x basis vector of it's transform.)
+#'
+#' C function prototype: \code{CP_EXPORT cpVect cpBodyGetRotation(const cpBody *body);}
+#'
+#' @param body [\code{cpBody *}]
+#'
+#' @return [\code{cpVect *}]
+#'
+#' @export
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+cpBodyGetRotation <- function(body) {
+  stopifnot(inherits(body, "cpBody"))
+  .Call("cpBodyGetRotation_", body)
 }