+++ src/cpBody.c
@@ -57,6 +57,21 @@ SEXP cpBodySetPosition_(SEXP body_, SEXP pos_)  {
 
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+// CP_EXPORT cpVect cpBodyGetCenterOfGravity(const cpBody *body);
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+SEXP cpBodyGetCenterOfGravity_(SEXP body_)  {
+  cpBody *body = isNull(body_) ? NULL : (cpBody *)R_ExternalPtrAddr(body_);
+  if (body == NULL) error("'cpBody * body' pointer is invalid/NULL");
+  cpVect * result = cpVect_obj_to_ptr(cpBodyGetCenterOfGravity(body));
+  SEXP result_ = PROTECT(R_MakeExternalPtr(result, R_NilValue, R_NilValue));
+  SET_CLASS(result_, mkString("cpVect"));
+  R_RegisterCFinalizer(result_, cpVect_finalizer);
+  UNPROTECT(1);
+  return result_;
+}
+
+
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // CP_EXPORT cpVect cpBodyGetVelocity(const cpBody *body);
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 SEXP cpBodyGetVelocity_(SEXP body_)  {
@@ -81,4 +96,83 @@ SEXP cpBodySetVelocity_(SEXP body_, SEXP velocity_)  {
   if (velocity == NULL) error("'cpVect * velocity' pointer is invalid/NULL");
   cpBodySetVelocity(body, *velocity);
   return R_NilValue;
+}
+
+
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+// CP_EXPORT cpVect cpBodyGetForce(const cpBody *body);
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+SEXP cpBodyGetForce_(SEXP body_)  {
+  cpBody *body = isNull(body_) ? NULL : (cpBody *)R_ExternalPtrAddr(body_);
+  if (body == NULL) error("'cpBody * body' pointer is invalid/NULL");
+  cpVect * result = cpVect_obj_to_ptr(cpBodyGetForce(body));
+  SEXP result_ = PROTECT(R_MakeExternalPtr(result, R_NilValue, R_NilValue));
+  SET_CLASS(result_, mkString("cpVect"));
+  R_RegisterCFinalizer(result_, cpVect_finalizer);
+  UNPROTECT(1);
+  return result_;
+}
+
+
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+// CP_EXPORT cpFloat cpBodyGetAngle(const cpBody *body);
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+SEXP cpBodyGetAngle_(SEXP body_)  {
+  cpBody *body = isNull(body_) ? NULL : (cpBody *)R_ExternalPtrAddr(body_);
+  if (body == NULL) error("'cpBody * body' pointer is invalid/NULL");
+  cpFloat result = cpBodyGetAngle(body);
+  SEXP result_ = PROTECT(ScalarReal(result));
+  UNPROTECT(1);
+  return result_;
+}
+
+
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+// CP_EXPORT void cpBodySetAngle(cpBody *body, cpFloat a);
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+SEXP cpBodySetAngle_(SEXP body_, SEXP a_)  {
+  cpBody *body = isNull(body_) ? NULL : (cpBody *)R_ExternalPtrAddr(body_);
+  if (body == NULL) error("'cpBody * body' pointer is invalid/NULL");
+  cpFloat a = isInteger(a_) ? (cpFloat)INTEGER(a_)[0] : (cpFloat)REAL(a_)[0];
+  cpBodySetAngle(body, a);
+  return R_NilValue;
+}
+
+
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+// CP_EXPORT void cpBodySetAngularVelocity(cpBody *body, cpFloat angularVelocity);
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+SEXP cpBodySetAngularVelocity_(SEXP body_, SEXP angularVelocity_)  {
+  cpBody *body = isNull(body_) ? NULL : (cpBody *)R_ExternalPtrAddr(body_);
+  if (body == NULL) error("'cpBody * body' pointer is invalid/NULL");
+  cpFloat angularVelocity = isInteger(angularVelocity_) ? (cpFloat)INTEGER(angularVelocity_)[0] : (cpFloat)REAL(angularVelocity_)[0];
+  cpBodySetAngularVelocity(body, angularVelocity);
+  return R_NilValue;
+}
+
+
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+// CP_EXPORT void cpBodySetTorque(cpBody *body, cpFloat torque);
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+SEXP cpBodySetTorque_(SEXP body_, SEXP torque_)  {
+  cpBody *body = isNull(body_) ? NULL : (cpBody *)R_ExternalPtrAddr(body_);
+  if (body == NULL) error("'cpBody * body' pointer is invalid/NULL");
+  cpFloat torque = isInteger(torque_) ? (cpFloat)INTEGER(torque_)[0] : (cpFloat)REAL(torque_)[0];
+  cpBodySetTorque(body, torque);
+  return R_NilValue;
+}
+
+
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+// CP_EXPORT cpVect cpBodyGetRotation(const cpBody *body);
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+SEXP cpBodyGetRotation_(SEXP body_)  {
+  cpBody *body = isNull(body_) ? NULL : (cpBody *)R_ExternalPtrAddr(body_);
+  if (body == NULL) error("'cpBody * body' pointer is invalid/NULL");
+  cpVect * result = cpVect_obj_to_ptr(cpBodyGetRotation(body));
+  SEXP result_ = PROTECT(R_MakeExternalPtr(result, R_NilValue, R_NilValue));
+  SET_CLASS(result_, mkString("cpVect"));
+  R_RegisterCFinalizer(result_, cpVect_finalizer);
+  UNPROTECT(1);
+  return result_;
 }