+++ NEWS.md
@@ -1,7 +1,13 @@
 
+# chipmunkcore 0.1.2
+
+* Features: Added box object
+* Internal: memory managment fix. remove shapes and bodies from a space before
+  freeing them.  This fixes a segfault.
+
 # chipmunkcore 0.1.1
 
-* Added elasticity controls
+* Feature: Added elasticity controls
 
 # chipmunkcore 0.1.0